pyproject.toml
[project]
name = "flywheel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.9.0",
    "langfuse>=4.0.0",
    "litellm>=1.82.1",
    "openai>=2.26.0",
    "openai-agents>=0.11.1",
    "python-dotenv>=1.2.2",
    "quixstreams>=3.23.4",
]
